Squash Adapter and Exchange
