+++ wayfinder_paths/adapters/hyperliquid_adapter/test_cancel_order.py
@@ -4,27 +4,28 @@ from unittest.mock import AsyncMock, patch
 import pytest
 
 from wayfinder_paths.adapters.hyperliquid_adapter.adapter import HyperliquidAdapter
-from wayfinder_paths.adapters.hyperliquid_adapter.exchange import Exchange
 
 
 class TestHyperliquidCancelOrder:
     @pytest.mark.asyncio
-    async def test_exchange_cancel_order_uses_int_oid(self):
+    async def test_adapter_cancel_order_uses_int_oid(self):
         with patch(
-            "wayfinder_paths.adapters.hyperliquid_adapter.exchange.get_info",
+            "wayfinder_paths.adapters.hyperliquid_adapter.adapter.get_info",
             return_value=SimpleNamespace(),
         ):
-            ex = Exchange(
-                sign_callback=AsyncMock(return_value="0x"),
-                signing_type="eip712",
+            adapter = HyperliquidAdapter(
+                config={},
+                sign_callback=AsyncMock(return_value="0x" + "00" * 65),
             )
-            ex.sign_and_broadcast_hypecore = AsyncMock(return_value={"status": "ok"})
+            adapter._sign_and_broadcast_hypecore = AsyncMock(
+                return_value={"status": "ok"}
+            )
 
-            await ex.cancel_order(
+            await adapter.cancel_order(
                 asset_id=10210, order_id=306356655993, address="0xabc"
             )
 
-            args, _ = ex.sign_and_broadcast_hypecore.await_args
+            args, _ = adapter._sign_and_broadcast_hypecore.await_args
             action = args[0]
             assert action["type"] == "cancel"
             assert action["cancels"][0]["a"] == 10210
@@ -33,24 +34,22 @@ class TestHyperliquidCancelOrder:
 
     @pytest.mark.asyncio
     async def test_adapter_cancel_order_parses_string_oid(self):
-        adapter = object.__new__(HyperliquidAdapter)
-        adapter._exchange = SimpleNamespace()
-        adapter._exchange.cancel_order = AsyncMock(return_value={"status": "ok"})
+        adapter = HyperliquidAdapter(config={})
+        adapter._sign_and_broadcast_hypecore = AsyncMock(return_value={"status": "ok"})
 
         ok, _ = await adapter.cancel_order(
             asset_id=10210, order_id="306356655993", address="0xabc"
         )
         assert ok is True
 
-        adapter._exchange.cancel_order.assert_awaited_once_with(
-            asset_id=10210, order_id=306356655993, address="0xabc"
-        )
+        args, _ = adapter._sign_and_broadcast_hypecore.await_args
+        action = args[0]
+        assert action["cancels"][0]["o"] == 306356655993
 
     @pytest.mark.asyncio
     async def test_adapter_cancel_order_rejects_bad_oid(self):
-        adapter = object.__new__(HyperliquidAdapter)
-        adapter._exchange = SimpleNamespace()
-        adapter._exchange.cancel_order = AsyncMock(return_value={"status": "ok"})
+        adapter = HyperliquidAdapter(config={})
+        adapter._sign_and_broadcast_hypecore = AsyncMock(return_value={"status": "ok"})
 
         ok, res = await adapter.cancel_order(
             asset_id=1, order_id="not-a-number", address="0xabc"
